+++ ntex-tls/src/schannel/connect.rs
@@ -1,7 +1,7 @@
 use std::io;
 
 use ntex_error::Error;
-use ntex_io::{Io, Layer};
+use ntex_io::{Filter, Io, Layer};
 use ntex_net::connect::{Address, Connect, ConnectError, Connector};
 use ntex_service::{Ctx, IntoService, Service, cfg::SharedCfg};
 use ntex_util::time::timeout_checked;
@@ -50,52 +50,66 @@ impl<A: Address> TlsConnector<Connector<A>> {
     }
 }
 
-impl<A: Address, S> Service<SharedCfg, Connect<A>> for TlsConnector<S>
-where
-    S: Service<SharedCfg, Connect<A>, Res = Io, Error = Error<ConnectError>>,
-{
-    type Res = Io<Layer<SchannelFilter>>;
-    type Error = Error<ConnectError>;
+impl<S> TlsConnector<S> {
+    /// Establish a TLS connection on top of an existing I/O stream.
+    pub async fn connect<F: Filter>(
+        &self,
+        io: Io<F>,
+        host: &str,
+        cfg: &SharedCfg,
+    ) -> Result<Io<Layer<SchannelFilter, F>>, Error<ConnectError>> {
+        let cfg = cfg.get::<TlsConfig>();
+        log::trace!("{}: TLS Handshake start for: {host:?} {io:?}", cfg.tag());
 
-    ntex_service::forward_ready!(SharedCfg, svc);
-    ntex_service::forward_shutdown!(SharedCfg, svc);
+        async {
+            match timeout_checked(
+                cfg.handshake_timeout(),
+                connect_io(io, host, self.config.clone()),
+            )
+            .await
+            {
+                Ok(Ok(io)) => {
+                    log::trace!("{}: TLS Handshake success: {host:?}", cfg.tag());
+                    Ok(io)
+                }
+                Ok(Err(e)) => {
+                    log::trace!("{}: TLS Handshake error: {e:?}", cfg.tag());
+                    Err(ConnectError::from(e).into())
+                }
+                Err(()) => {
+                    log::trace!("{}: TLS Handshake timeout", cfg.tag());
+                    Err(ConnectError::from(io::Error::new(
+                        io::ErrorKind::TimedOut,
+                        "TLS Handshake timeout",
+                    ))
+                    .into())
+                }
+            }
+        }
+        .await
+        .map_err(|e: Error<_>| e.set_service(cfg.service()))
+    }
+}
+
+impl<F: Filter, A: Address, S> Service<SharedCfg, Connect<A>> for TlsConnector<S>
+where
+    S: Service<SharedCfg, Connect<A>, Res = Io<F>, Error = Error<ConnectError>>,
+{
+    type Res = Io<Layer<SchannelFilter, F>>;
+    type Error = Error<ConnectError>;
 
     async fn call(
         &self,
-        message: Connect<A>,
+        req: Connect<A>,
         ctx: Ctx<'_, Self, SharedCfg>,
     ) -> Result<Self::Res, Self::Error> {
-        let cfg = ctx.st().get::<TlsConfig>();
-        let host = message.host().split(':').next().unwrap().to_string();
-
-        let io = ctx.call(&self.svc, message).await?;
-        let tag = io.tag();
-        log::trace!("{tag}: TLS Handshake start for: {host:?}");
-
-        match timeout_checked(
-            cfg.handshake_timeout(),
-            connect_io(io, &host, self.config.clone()),
-        )
-        .await
-        {
-            Ok(Ok(io)) => {
-                log::trace!("{tag}: TLS Handshake success: {host:?}");
-                Ok(io)
-            }
-            Ok(Err(e)) => {
-                log::trace!("{tag}: TLS Handshake error: {e:?}");
-                Err(ConnectError::from(e).into())
-            }
-            Err(()) => {
-                log::trace!("{tag}: TLS Handshake timeout");
-                Err(ConnectError::from(io::Error::new(
-                    io::ErrorKind::TimedOut,
-                    "TLS Handshake timeout",
-                ))
-                .into())
-            }
-        }
+        let host = req.host().split(':').next().unwrap().to_string();
+        let io = ctx.call(&self.svc, req).await?;
+        self.connect(io, &host, ctx.st()).await
     }
+
+    ntex_service::forward_ready!(SharedCfg, svc);
+    ntex_service::forward_shutdown!(SharedCfg, svc);
 }
 
 #[cfg(test)]
@@ -112,7 +126,7 @@ mod tests {
 
         let svc: TlsConnector<Connector<&'static str>> = TlsConnector::new();
         assert!(format!("{svc:?}").contains("TlsConnector"));
-        let srv = Pipeline::new(svc);
+        let srv = Pipeline::with(SharedCfg::default(), svc);
         assert!(srv.ready().await.is_ok());
         let result = srv
             .call(Connect::new("").set_addr(Some(server.addr())))