+++ ntex/src/web/error_default.rs
@@ -6,7 +6,9 @@ use serde_json::error::Error as JsonError;
 use serde_urlencoded::ser::Error as FormError;
 
 use crate::client;
-use crate::http::{self, StatusCode, header};
+#[cfg(feature = "ws")]
+use crate::http::header;
+use crate::http::{self, StatusCode};
 use crate::util::timeout::TimeoutError;
 #[cfg(feature = "ws")]
 use crate::ws::error::HandshakeError;
@@ -268,6 +270,7 @@ impl WebResponseError<WebError> for HandshakeError {
     }
 }
 
+#[cfg(feature = "ws")]
 impl From<HandshakeError> for WebError {
     fn from(err: HandshakeError) -> Self {
         Self::new(err)